move RandomGenerator class in its own file
